refactor (cmdserver): Add new command data and change handler.
Updated command data structs to include new fields that will be
available starting with Diatheke v3. These fields include session
and model IDs and application-defined metadata.

Also updated the handler function signature to pass in the command
output struct so that it is no longer the handler's responsibility
to instantiate it. Instead all the handler needs to do is set
values on the output data. Also changed the signature so that it
passes in all the new data fields instead of only the command
parameters.